Who passed the lesson, who left?
You are working in the IT department of X University and you are asked which students in computer engineering should repeat this period.  At X university, students whose average is below the average of the department have to repeat that semester.  All students' names, surnames, and the grades they receive in the lessons reach you in a text file called “student.txt”.  In that period, only 3 courses were opened in computer engineering and all students take these 3 courses.  Identify the structure named struct student whose members are first name (string / string), last name (string / string), notes (3-element float array) and average (float).  Transfer the records in the file "student.txt" into a string of up to 100 elements of the type struct student.  After calculating the averages, you have to write the names and grade averages of the repeating students into the file called “repeat.txt”.  Write a full C program that performs the above process.  NOTE: The maximum student capacity of the university is 100. You can assume that names and surnames have a maximum of 20 characters each.